+++ src/java/org/smoothbuild/app/RunnableCommand.java
@@ -1,8 +1,0 @@
-package org.smoothbuild.app;
-
-public interface RunnableCommand {
-  /**
-   * @return true on success on false when problems occurred
-   */
-  public boolean runCommand();
-}