+++ src/java/org/smoothbuild/cli/handle/Handler.java
@@ -1,0 +1,7 @@
+package org.smoothbuild.cli.handle;
+
+import net.sourceforge.argparse4j.inf.Namespace;
+
+public interface Handler {
+  public boolean run(Namespace namespace);
+}